scan_processing_service/requirements.txt
open3d
alembic
asgi_lifespan
argon2_cffi
asyncio
asyncpg
psycopg2-binary
bcrypt
black
colorama
email-validator
fastapi
greenlet
httpx
isort
loguru
mypy
passlib
pathlib
pillow
pre-commit
python-dotenv
python-decouple
python-jose
python-multipart
python-slugify
pytest
pytest-asyncio
pytest-cov
pytest-xdist
SQLAlchemy
trio
uvicorn
pydantic_settings
password_strength
numpy
matplotlib
shapely
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
google==3.0.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
requests==2.31.0
rsa==4.9
soupsieve==2.5
uritemplate==4.1.1
urllib3==2.1.0